moved prepareAreaCodeProximityMap to scripts/prepareAreaCodeProximityMap
